+++ R/filter_by_shape_id.R
@@ -159,7 +159,7 @@ filter_by_shape_id <- function(gtfs, shape_id, keep = TRUE) {
     # 'agency' (agency_id, that comes both from routes and fare_attributes)
 
     if (gtfsio::check_field_exists(gtfs, "agency", "agency_id") &&
-        length(relevant_agencies) > 0) {
+        exists("relevant_agencies")) {
 
       # keeping only unique agency_ids from relevant_agencies, since they may
       # come from two different sources